Tests/LibWeb/Text/input/navigation/hashchange-event.html

<script src="../include.js"></script>
<iframe src="../../data/iframe-hashchange-event.html"></iframe>
<script>
    asyncTest(done => {
        let messageCount = 0;
        window.addEventListener("message", event => {
            println(event.data);
            messageCount++;
            if (messageCount == 2)
                done();
        });
    });
</script>
