Tests/LibWeb/Text/expected/navigation/hashchange-event.txt

hashchange oldURL.hash= newURL.hash=#test
hashchange oldURL.hash=#test newURL.hash=#done
